refac: move some files to `zk/plonk/constraint_system`
- `Challenge`
- `Circuit`
- `Constraint`
- `ConstraintSystem`
- `Gate`
- `Query`
- `Rotation`
- `SelectorXXX`
- `VirtualCellXXX`
