+++ tachyon/zk/expressions/expression_stringifier.h
@@ -18,9 +18,9 @@
 #include "tachyon/zk/expressions/scaled_expression.h"
 #include "tachyon/zk/expressions/selector_expression.h"
 #include "tachyon/zk/expressions/sum_expression.h"
-#include "tachyon/zk/plonk/circuit/challenge_stringifier.h"
 #include "tachyon/zk/plonk/circuit/phase_stringifier.h"
-#include "tachyon/zk/plonk/circuit/rotation_stringifier.h"
+#include "tachyon/zk/plonk/constraint_system/challenge_stringifier.h"
+#include "tachyon/zk/plonk/constraint_system/rotation_stringifier.h"
 
 namespace tachyon::base::internal {
