+++ tachyon/zk/lookup/BUILD.bazel
@@ -35,7 +35,7 @@ tachyon_cc_library(
         "//tachyon/zk/base/entities:prover_base",
         "//tachyon/zk/expressions/evaluator:simple_evaluator",
         "//tachyon/zk/lookup:permute_expression_pair",
-        "//tachyon/zk/plonk/circuit:rotation",
+        "//tachyon/zk/plonk/constraint_system:rotation",
         "//tachyon/zk/plonk/permutation:grand_product_argument",
     ],
 )
