--- tachyon/zk/plonk/BUILD.bazel
@@ -1,34 +1,0 @@
-load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
-
-package(default_visibility = ["//visibility:public"])
-
-tachyon_cc_library(
-    name = "constraint_system",
-    hdrs = ["constraint_system.h"],
-    deps = [
-        "//tachyon/base/containers:container_util",
-        "//tachyon/base/containers:contains",
-        "//tachyon/base/functional:callback",
-        "//tachyon/zk/base:row_index",
-        "//tachyon/zk/expressions/evaluator:simple_selector_finder",
-        "//tachyon/zk/lookup:lookup_argument",
-        "//tachyon/zk/plonk/circuit:constraint",
-        "//tachyon/zk/plonk/circuit:gate",
-        "//tachyon/zk/plonk/circuit:lookup_table_column",
-        "//tachyon/zk/plonk/circuit:query",
-        "//tachyon/zk/plonk/circuit:selector_compressor",
-        "//tachyon/zk/plonk/circuit:virtual_cells",
-        "//tachyon/zk/plonk/permutation:permutation_argument",
-        "//tachyon/zk/plonk/permutation:permutation_utils",
-        "@com_google_absl//absl/container:flat_hash_map",
-    ],
-)
-
-tachyon_cc_unittest(
-    name = "plonk_unittests",
-    srcs = ["constraint_system_unittest.cc"],
-    deps = [
-        ":constraint_system",
-        "//tachyon/math/finite_fields/test:gf7",
-    ],
-)
