+++ tachyon/zk/plonk/circuit/BUILD.bazel
@@ -16,13 +16,13 @@ tachyon_cc_library(
     name = "assignment",
     hdrs = ["assignment.h"],
     deps = [
-        ":challenge",
         ":column_key",
-        ":selector",
         "//tachyon/base/functional:callback",
         "//tachyon/math/base:rational_field",
         "//tachyon/zk/base:row_index",
         "//tachyon/zk/base:value",
+        "//tachyon/zk/plonk/constraint_system:challenge",
+        "//tachyon/zk/plonk/constraint_system:selector",
     ],
 )
 
@@ -34,33 +34,6 @@ tachyon_cc_library(
         "//tachyon:export",
         "//tachyon/zk/base:row_index",
         "@com_google_absl//absl/strings",
-    ],
-)
-
-tachyon_cc_library(
-    name = "challenge",
-    hdrs = ["challenge.h"],
-    deps = [
-        ":phase",
-        "@com_google_absl//absl/strings",
-    ],
-)
-
-tachyon_cc_library(
-    name = "challenge_stringifier",
-    hdrs = ["challenge_stringifier.h"],
-    deps = [
-        ":challenge",
-        ":phase_stringifier",
-    ],
-)
-
-tachyon_cc_library(
-    name = "circuit",
-    hdrs = ["circuit.h"],
-    deps = [
-        ":layouter",
-        "//tachyon/zk/plonk/circuit/floor_planner",
     ],
 )
 
@@ -101,45 +74,18 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "constraint",
-    hdrs = ["constraint.h"],
-    deps = ["//tachyon/zk/expressions:expression"],
-)
-
-tachyon_cc_library(
-    name = "exclusion_matrix",
-    hdrs = ["exclusion_matrix.h"],
-    deps = [
-        ":selector_description",
-        "//tachyon:export",
-        "//tachyon/base/containers:container_util",
-    ],
-)
-
-tachyon_cc_library(
-    name = "gate",
-    hdrs = ["gate.h"],
-    deps = [
-        ":selector",
-        ":virtual_cell",
-        "//tachyon/zk/expressions:expression",
-        "@com_google_absl//absl/strings",
-    ],
-)
-
-tachyon_cc_library(
     name = "layouter",
     hdrs = [
         "layouter.h",
         "namespaced_layouter.h",
     ],
     deps = [
-        ":challenge",
         ":lookup_table",
         ":region",
         "//tachyon/base/functional:callback",
         "//tachyon/zk/base:row_index",
-        "//tachyon/zk/plonk:constraint_system",
+        "//tachyon/zk/plonk/constraint_system",
+        "//tachyon/zk/plonk/constraint_system:challenge",
     ],
 )
 
@@ -188,25 +134,6 @@ tachyon_cc_library(
 )
 
 tachyon_cc_library(
-    name = "query",
-    hdrs = ["query.h"],
-    deps = [
-        ":column_key",
-        ":rotation",
-    ],
-)
-
-tachyon_cc_library(
-    name = "query_stringifier",
-    hdrs = ["query_stringifier.h"],
-    deps = [
-        ":column_key_stringifier",
-        ":query",
-        ":rotation_stringifier",
-    ],
-)
-
-tachyon_cc_library(
     name = "ref_table",
     hdrs = ["ref_table.h"],
     deps = [":table_base"],
@@ -219,8 +146,8 @@ tachyon_cc_library(
         ":assigned_cell",
         ":column_key",
         ":region_layouter",
-        ":selector",
         "//tachyon/base/functional:callback",
+        "//tachyon/zk/plonk/constraint_system:selector",
     ],
 )
 
@@ -229,7 +156,7 @@ tachyon_cc_library(
     hdrs = ["region_column.h"],
     deps = [
         ":column_key",
-        ":selector",
+        "//tachyon/zk/plonk/constraint_system:selector",
     ],
 )
 
@@ -239,11 +166,11 @@ tachyon_cc_library(
     deps = [
         ":assigned_cell",
         ":column_key",
-        ":selector",
         "//tachyon/base/functional:callback",
         "//tachyon/math/base:rational_field",
         "//tachyon/zk/base:row_index",
         "//tachyon/zk/base:value",
+        "//tachyon/zk/plonk/constraint_system:selector",
     ],
 )
 
@@ -254,68 +181,6 @@ tachyon_cc_library(
         ":region",
         ":region_column",
         "@com_google_absl//absl/container:flat_hash_set",
-    ],
-)
-
-tachyon_cc_library(
-    name = "rotation",
-    hdrs = ["rotation.h"],
-    deps = [
-        "//tachyon/base:logging",
-        "//tachyon/base/numerics:checked_math",
-        "//tachyon/base/strings:string_number_conversions",
-        "//tachyon/math/polynomials/univariate:univariate_evaluation_domain",
-        "//tachyon/zk/base:row_index",
-    ],
-)
-
-tachyon_cc_library(
-    name = "rotation_stringifier",
-    hdrs = ["rotation_stringifier.h"],
-    deps = [
-        ":rotation",
-        "//tachyon/base/strings:rust_stringifier",
-    ],
-)
-
-tachyon_cc_library(
-    name = "selector",
-    hdrs = ["selector.h"],
-    deps = [
-        "//tachyon:export",
-        "//tachyon/zk/base:row_index",
-        "@com_google_absl//absl/hash",
-        "@com_google_absl//absl/strings",
-    ],
-)
-
-tachyon_cc_library(
-    name = "selector_assignment",
-    hdrs = ["selector_assignment.h"],
-    deps = ["//tachyon/zk/expressions:expression"],
-)
-
-tachyon_cc_library(
-    name = "selector_compressor",
-    hdrs = ["selector_compressor.h"],
-    deps = [
-        ":exclusion_matrix",
-        ":selector_assignment",
-        ":selector_description",
-        "//tachyon/base/containers:container_util",
-        "//tachyon/base/containers:cxx20_erase",
-        "//tachyon/base/functional:callback",
-        "//tachyon/zk/expressions:expression_factory",
-        "@com_google_googletest//:gtest_prod",
-    ],
-)
-
-tachyon_cc_library(
-    name = "selector_description",
-    hdrs = ["selector_description.h"],
-    deps = [
-        "//tachyon:export",
-        "//tachyon/base/containers:adapters",
     ],
 )
 
@@ -332,52 +197,23 @@ tachyon_cc_library(
     ],
 )
 
-tachyon_cc_library(
-    name = "virtual_cell",
-    hdrs = ["virtual_cell.h"],
-    deps = [
-        ":column_key",
-        ":rotation",
-        "//tachyon:export",
-    ],
-)
-
-tachyon_cc_library(
-    name = "virtual_cells",
-    hdrs = ["virtual_cells.h"],
-    deps = [
-        ":selector",
-        ":virtual_cell",
-        "//tachyon/zk/expressions:expression_factory",
-    ],
-)
-
 tachyon_cc_unittest(
     name = "circuit_unittests",
     srcs = [
         "column_key_unittest.cc",
-        "exclusion_matrix_unittest.cc",
         "lookup_table_column_unittest.cc",
         "phase_unittest.cc",
         "ref_table_unittest.cc",
         "region_column_unittest.cc",
-        "rotation_unittest.cc",
-        "selector_compressor_unittest.cc",
-        "selector_description_unittest.cc",
-        "selector_unittest.cc",
     ],
     deps = [
         ":column_key",
-        ":exclusion_matrix",
         ":lookup_table_column",
         ":ref_table",
         ":region_column",
-        ":rotation",
-        ":selector_compressor",
-        ":selector_description",
         "//tachyon/math/elliptic_curves/bn/bn254:g1",
         "//tachyon/math/finite_fields/test:gf7",
-        "//tachyon/math/polynomials/univariate:univariate_polynomial",
+        "//tachyon/math/polynomials/univariate:univariate_evaluations",
         "@com_google_absl//absl/hash:hash_testing",
     ],
 )
