+++ tachyon/zk/plonk/circuit/examples/simple_circuit.h
@@ -12,7 +12,7 @@
 #include <utility>
 #include <vector>
 
-#include "tachyon/zk/plonk/circuit/circuit.h"
+#include "tachyon/zk/plonk/constraint_system/circuit.h"
 
 namespace tachyon::zk {
