+++ tachyon/zk/plonk/circuit/region_column.h
@@ -14,7 +14,7 @@
 
 #include "tachyon/export.h"
 #include "tachyon/zk/plonk/circuit/column_key.h"
-#include "tachyon/zk/plonk/circuit/selector.h"
+#include "tachyon/zk/plonk/constraint_system/selector.h"
 
 namespace tachyon::zk {
