+++ tachyon/zk/plonk/constraint_system/BUILD.bazel
@@ -1,0 +1,197 @@
+load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
+
+package(default_visibility = ["//visibility:public"])
+
+tachyon_cc_library(
+    name = "challenge",
+    hdrs = ["challenge.h"],
+    deps = [
+        "//tachyon/zk/plonk/circuit:phase",
+        "@com_google_absl//absl/strings",
+    ],
+)
+
+tachyon_cc_library(
+    name = "challenge_stringifier",
+    hdrs = ["challenge_stringifier.h"],
+    deps = [
+        ":challenge",
+        "//tachyon/zk/plonk/circuit:phase_stringifier",
+    ],
+)
+
+tachyon_cc_library(
+    name = "circuit",
+    hdrs = ["circuit.h"],
+    deps = [
+        "//tachyon/zk/plonk/circuit:layouter",
+        "//tachyon/zk/plonk/circuit/floor_planner",
+    ],
+)
+
+tachyon_cc_library(
+    name = "constraint",
+    hdrs = ["constraint.h"],
+    deps = ["//tachyon/zk/expressions:expression"],
+)
+
+tachyon_cc_library(
+    name = "constraint_system",
+    hdrs = ["constraint_system.h"],
+    deps = [
+        "//tachyon/base/containers:container_util",
+        "//tachyon/base/containers:contains",
+        "//tachyon/base/functional:callback",
+        "//tachyon/zk/base:row_index",
+        "//tachyon/zk/expressions/evaluator:simple_selector_finder",
+        "//tachyon/zk/lookup:lookup_argument",
+        "//tachyon/zk/plonk/circuit:lookup_table_column",
+        "//tachyon/zk/plonk/constraint_system:constraint",
+        "//tachyon/zk/plonk/constraint_system:gate",
+        "//tachyon/zk/plonk/constraint_system:query",
+        "//tachyon/zk/plonk/constraint_system:selector_compressor",
+        "//tachyon/zk/plonk/constraint_system:virtual_cells",
+        "//tachyon/zk/plonk/permutation:permutation_argument",
+        "//tachyon/zk/plonk/permutation:permutation_utils",
+        "@com_google_absl//absl/container:flat_hash_map",
+    ],
+)
+
+tachyon_cc_library(
+    name = "exclusion_matrix",
+    hdrs = ["exclusion_matrix.h"],
+    deps = [
+        ":selector_description",
+        "//tachyon:export",
+        "//tachyon/base/containers:container_util",
+    ],
+)
+
+tachyon_cc_library(
+    name = "gate",
+    hdrs = ["gate.h"],
+    deps = [
+        ":selector",
+        ":virtual_cell",
+        "//tachyon/zk/expressions:expression",
+        "@com_google_absl//absl/strings",
+    ],
+)
+
+tachyon_cc_library(
+    name = "query",
+    hdrs = ["query.h"],
+    deps = [
+        ":rotation",
+        "//tachyon/zk/plonk/circuit:column_key",
+    ],
+)
+
+tachyon_cc_library(
+    name = "query_stringifier",
+    hdrs = ["query_stringifier.h"],
+    deps = [
+        ":query",
+        ":rotation_stringifier",
+        "//tachyon/zk/plonk/circuit:column_key_stringifier",
+    ],
+)
+
+tachyon_cc_library(
+    name = "rotation",
+    hdrs = ["rotation.h"],
+    deps = [
+        "//tachyon/base:logging",
+        "//tachyon/base/numerics:checked_math",
+        "//tachyon/base/strings:string_number_conversions",
+        "//tachyon/math/polynomials/univariate:univariate_evaluation_domain",
+        "//tachyon/zk/base:row_index",
+    ],
+)
+
+tachyon_cc_library(
+    name = "rotation_stringifier",
+    hdrs = ["rotation_stringifier.h"],
+    deps = [
+        ":rotation",
+        "//tachyon/base/strings:rust_stringifier",
+    ],
+)
+
+tachyon_cc_library(
+    name = "selector",
+    hdrs = ["selector.h"],
+    deps = [
+        "//tachyon:export",
+        "//tachyon/zk/base:row_index",
+        "@com_google_absl//absl/hash",
+        "@com_google_absl//absl/strings",
+    ],
+)
+
+tachyon_cc_library(
+    name = "selector_assignment",
+    hdrs = ["selector_assignment.h"],
+    deps = ["//tachyon/zk/expressions:expression"],
+)
+
+tachyon_cc_library(
+    name = "selector_compressor",
+    hdrs = ["selector_compressor.h"],
+    deps = [
+        ":exclusion_matrix",
+        ":selector_assignment",
+        ":selector_description",
+        "//tachyon/base/containers:container_util",
+        "//tachyon/base/containers:cxx20_erase",
+        "//tachyon/base/functional:callback",
+        "//tachyon/zk/expressions:expression_factory",
+        "@com_google_googletest//:gtest_prod",
+    ],
+)
+
+tachyon_cc_library(
+    name = "selector_description",
+    hdrs = ["selector_description.h"],
+    deps = [
+        "//tachyon:export",
+        "//tachyon/base/containers:adapters",
+    ],
+)
+
+tachyon_cc_library(
+    name = "virtual_cell",
+    hdrs = ["virtual_cell.h"],
+    deps = [
+        ":rotation",
+        "//tachyon:export",
+        "//tachyon/zk/plonk/circuit:column_key",
+    ],
+)
+
+tachyon_cc_library(
+    name = "virtual_cells",
+    hdrs = ["virtual_cells.h"],
+    deps = [
+        ":selector",
+        ":virtual_cell",
+        "//tachyon/zk/expressions:expression_factory",
+    ],
+)
+
+tachyon_cc_unittest(
+    name = "plonk_unittests",
+    srcs = [
+        "constraint_system_unittest.cc",
+        "exclusion_matrix_unittest.cc",
+        "rotation_unittest.cc",
+        "selector_compressor_unittest.cc",
+        "selector_description_unittest.cc",
+        "selector_unittest.cc",
+    ],
+    deps = [
+        ":constraint_system",
+        "//tachyon/math/finite_fields/test:gf7",
+        "@com_google_absl//absl/hash:hash_testing",
+    ],
+)
