+++ tachyon/zk/plonk/constraint_system/challenge_stringifier.h
@@ -4,13 +4,13 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_CHALLENGE_STRINGIFIER_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_CHALLENGE_STRINGIFIER_H_
+#ifndef TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CHALLENGE_STRINGIFIER_H_
+#define TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CHALLENGE_STRINGIFIER_H_
 
 #include <ostream>
 
 #include "tachyon/base/strings/rust_stringifier.h"
-#include "tachyon/zk/plonk/circuit/challenge.h"
+#include "tachyon/zk/plonk/constraint_system/challenge.h"
 
 namespace tachyon::base::internal {
 
@@ -28,4 +28,4 @@ class RustDebugStringifier<zk::Challenge> {
 
 }  // namespace tachyon::base::internal
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_CHALLENGE_STRINGIFIER_H_
+#endif  // TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CHALLENGE_STRINGIFIER_H_
