+++ tachyon/zk/plonk/constraint_system/circuit.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_CIRCUIT_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_CIRCUIT_H_
+#ifndef TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CIRCUIT_H_
+#define TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CIRCUIT_H_
 
 #include <memory>
 
@@ -28,4 +28,4 @@ class Circuit {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_CIRCUIT_H_
+#endif  // TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CIRCUIT_H_
