+++ tachyon/zk/plonk/constraint_system/constraint.h
@@ -1,5 +1,5 @@
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_CONSTRAINT_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_CONSTRAINT_H_
+#ifndef TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CONSTRAINT_H_
+#define TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CONSTRAINT_H_
 
 #include <memory>
 #include <string>
@@ -45,4 +45,4 @@ class Constraint {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_CONSTRAINT_H_
+#endif  // TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_CONSTRAINT_H_
