+++ tachyon/zk/plonk/constraint_system/selector_assignment.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_SELECTOR_ASSIGNMENT_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_SELECTOR_ASSIGNMENT_H_
+#ifndef TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_SELECTOR_ASSIGNMENT_H_
+#define TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_SELECTOR_ASSIGNMENT_H_
 
 #include <memory>
 #include <utility>
@@ -41,4 +41,4 @@ class SelectorAssignment {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_SELECTOR_ASSIGNMENT_H_
+#endif  // TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_SELECTOR_ASSIGNMENT_H_
