+++ tachyon/zk/plonk/constraint_system/selector_description.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_SELECTOR_DESCRIPTION_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_SELECTOR_DESCRIPTION_H_
+#ifndef TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_SELECTOR_DESCRIPTION_H_
+#define TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_SELECTOR_DESCRIPTION_H_
 
 #include <utility>
 #include <vector>
@@ -57,4 +57,4 @@ class TACHYON_EXPORT SelectorDescription {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_SELECTOR_DESCRIPTION_H_
+#endif  // TACHYON_ZK_PLONK_CONSTRAINT_SYSTEM_SELECTOR_DESCRIPTION_H_
