+++ tachyon/zk/plonk/halo2/BUILD.bazel
@@ -38,9 +38,9 @@ tachyon_cc_library(
         "//tachyon/zk/base:point_set",
         "//tachyon/zk/base/entities:prover_base",
         "//tachyon/zk/lookup:lookup_argument_runner",
-        "//tachyon/zk/plonk:constraint_system",
-        "//tachyon/zk/plonk/circuit:query",
         "//tachyon/zk/plonk/circuit:ref_table",
+        "//tachyon/zk/plonk/constraint_system",
+        "//tachyon/zk/plonk/constraint_system:query",
         "//tachyon/zk/plonk/keys:proving_key",
         "//tachyon/zk/plonk/permutation:permutation_argument_runner",
         "@com_google_absl//absl/types:span",
@@ -71,9 +71,9 @@ tachyon_cc_library(
     deps = [
         ":pinned_gates",
         "//tachyon/zk/lookup:lookup_argument_stringifier",
-        "//tachyon/zk/plonk:constraint_system",
         "//tachyon/zk/plonk/circuit:phase_stringifier",
-        "//tachyon/zk/plonk/circuit:query_stringifier",
+        "//tachyon/zk/plonk/constraint_system",
+        "//tachyon/zk/plonk/constraint_system:query_stringifier",
         "//tachyon/zk/plonk/permutation:permutation_argument_stringifier",
     ],
 )
@@ -218,7 +218,7 @@ tachyon_cc_library(
         ":witness_collection",
         "//tachyon/base/containers:container_util",
         "//tachyon/zk/base/entities:prover_base",
-        "//tachyon/zk/plonk:constraint_system",
+        "//tachyon/zk/plonk/constraint_system",
     ],
 )
