+++ tachyon/zk/plonk/halo2/argument_util.h
@@ -14,9 +14,9 @@
 #include "tachyon/zk/base/entities/prover_base.h"
 #include "tachyon/zk/base/point_set.h"
 #include "tachyon/zk/lookup/lookup_argument_runner.h"
-#include "tachyon/zk/plonk/circuit/query.h"
 #include "tachyon/zk/plonk/circuit/ref_table.h"
-#include "tachyon/zk/plonk/constraint_system.h"
+#include "tachyon/zk/plonk/constraint_system/constraint_system.h"
+#include "tachyon/zk/plonk/constraint_system/query.h"
 #include "tachyon/zk/plonk/halo2/step_returns.h"
 #include "tachyon/zk/plonk/keys/proving_key.h"
 #include "tachyon/zk/plonk/permutation/permutation_argument_runner.h"
