+++ tachyon/zk/plonk/permutation/BUILD.bazel
@@ -60,7 +60,7 @@ tachyon_cc_library(
         "//tachyon/crypto/commitments:polynomial_openings",
         "//tachyon/zk/base:point_set",
         "//tachyon/zk/base/entities:prover_base",
-        "//tachyon/zk/plonk/circuit:rotation",
+        "//tachyon/zk/plonk/constraint_system:rotation",
     ],
 )
 
@@ -129,7 +129,7 @@ tachyon_cc_library(
         "//tachyon/base:logging",
         "//tachyon/base/containers:adapters",
         "//tachyon/crypto/commitments:polynomial_openings",
-        "//tachyon/zk/plonk:constraint_system",
+        "//tachyon/zk/plonk/constraint_system",
     ],
 )
