+++ tachyon/zk/plonk/vanishing/BUILD.bazel
@@ -27,7 +27,7 @@ tachyon_cc_library(
         "//tachyon/zk/plonk/circuit:column_key",
         "//tachyon/zk/plonk/circuit:owned_table",
         "//tachyon/zk/plonk/circuit:ref_table",
-        "//tachyon/zk/plonk/circuit:rotation",
+        "//tachyon/zk/plonk/constraint_system:rotation",
         "//tachyon/zk/plonk/permutation:permutation_committed",
         "//tachyon/zk/plonk/permutation:unpermuted_table",
         "@com_google_absl//absl/types:span",
@@ -94,7 +94,7 @@ tachyon_cc_library(
         ":circuit_polynomial_builder",
         ":graph_evaluator",
         "//tachyon/base/containers:container_util",
-        "//tachyon/zk/plonk:constraint_system",
+        "//tachyon/zk/plonk/constraint_system",
     ],
 )
 
@@ -173,10 +173,10 @@ tachyon_cc_unittest(
         "//tachyon/zk/base/entities:verifier_base",
         "//tachyon/zk/expressions:expression_factory",
         "//tachyon/zk/expressions/evaluator/test:evaluator_test",
-        "//tachyon/zk/plonk:constraint_system",
         "//tachyon/zk/plonk/circuit/examples:circuit_test",
         "//tachyon/zk/plonk/circuit/examples:simple_circuit",
         "//tachyon/zk/plonk/circuit/floor_planner:simple_floor_planner",
+        "//tachyon/zk/plonk/constraint_system",
         "//tachyon/zk/plonk/halo2:pinned_verifying_key",
         "//tachyon/zk/plonk/halo2:prover_test",
         "//tachyon/zk/plonk/keys:proving_key",
