+++ vendors/halo2/BUILD.bazel
@@ -227,8 +227,8 @@ tachyon_cc_library(
         "//tachyon/math/polynomials/univariate:univariate_polynomial",
         "//tachyon/zk/base/commitments:shplonk_extension",
         "//tachyon/zk/plonk/circuit:column_key",
-        "//tachyon/zk/plonk/circuit:gate",
         "//tachyon/zk/plonk/circuit:phase",
+        "//tachyon/zk/plonk/constraint_system:gate",
         "//tachyon/zk/plonk/halo2:pinned_verifying_key",
         "//tachyon/zk/plonk/keys:proving_key",
         "//tachyon/zk/plonk/permutation:permutation_argument",
